separted metrics reporting and Performance sampling into two traits to improve clarity
