+++ core/src/test/scala/kanaloa/reactive/dispatcher/PerformanceSamplerSpec.scala
@@ -1,37 +1,38 @@
-package kanaloa.reactive.dispatcher.metrics
+package kanaloa.reactive.dispatcher
 
-import akka.actor.{ActorSystem, ActorRef}
-import kanaloa.reactive.dispatcher.SpecWithActorSystem
+import akka.actor.{ActorRef, ActorSystem}
+import kanaloa.reactive.dispatcher.PerformanceSampler._
 import kanaloa.reactive.dispatcher.metrics.Metric._
-import kanaloa.reactive.dispatcher.metrics.MetricsCollector._
+import kanaloa.reactive.dispatcher.metrics.{MetricsCollector, Reporter}
+import org.mockito.Mockito._
 import org.scalatest.concurrent.Eventually
 import org.scalatest.mock.MockitoSugar
-import concurrent.duration._
-import org.mockito.Mockito._
 
-class MetricsCollectorSpec extends SpecWithActorSystem with MockitoSugar with Eventually {
+import scala.concurrent.duration._
+
+class PerformanceSamplerSpec extends SpecWithActorSystem with MockitoSugar with Eventually {
   val waitDuration = 30.milliseconds
 
-  def initMetricsCollector(minSampleDurationRatio: Double = 0)(implicit system: ActorSystem): ActorRef = {
-    val mc = system.actorOf(MetricsCollector.props(None, MetricsCollectorSettings(sampleRate = waitDuration / 2, minSampleDurationRatio = minSampleDurationRatio)))
-    mc ! fullyUtilizedResult //set it in the busy mode
-    mc ! PoolSize(10)
-    mc ! Subscribe(self)
-    mc
+  def initPerformanceSampler(minSampleDurationRatio: Double = 0)(implicit system: ActorSystem): ActorRef = {
+    val ps = system.actorOf(MetricsCollector.props(None, PerformanceSamplerSettings(sampleRate = waitDuration / 2, minSampleDurationRatio = minSampleDurationRatio)))
+    ps ! fullyUtilizedResult //set it in the busy mode
+    ps ! PoolSize(10)
+    ps ! Subscribe(self)
+    ps
   }
   val partialUtilizedResult: DispatchResult = DispatchResult(1, 0)
   val fullyUtilizedResult: DispatchResult = DispatchResult(0, 2)
 
-  "MetricsCollector" should {
+  "PerformanceSampler" should {
     "send Samples periodically" in {
-      val mc = initMetricsCollector()
-      mc ! WorkCompleted(1.millisecond)
-      mc ! WorkCompleted(1.millisecond)
+      val ps = initPerformanceSampler()
+      ps ! WorkCompleted(1.millisecond)
+      ps ! WorkCompleted(1.millisecond)
 
       val sample1 = expectMsgType[Sample]
       sample1.workDone should be(2)
 
-      mc ! WorkCompleted(1.millisecond)
+      ps ! WorkCompleted(1.millisecond)
 
       val sample2 = expectMsgType[Sample]
       sample2.workDone should be(1)
@@ -41,51 +42,51 @@ class MetricsCollectorSpec extends SpecWithActorSystem with MockitoSugar with Ev
     }
 
     "ignore metrics when pool isn't fully occupied" in {
-      val mc = initMetricsCollector()
-      mc ! partialUtilizedResult
+      val ps = initPerformanceSampler()
+      ps ! partialUtilizedResult
       expectMsgType[PartialUtilization].numOfBusyWorkers should be(9)
 
-      mc ! WorkCompleted(1.millisecond)
-      mc ! WorkCompleted(1.millisecond)
+      ps ! WorkCompleted(1.millisecond)
+      ps ! WorkCompleted(1.millisecond)
 
       expectNoMsg(waitDuration)
 
     }
 
     "ignore Work timeout but include failed Work " in {
-      val mc = initMetricsCollector()
-      mc ! WorkTimedOut
+      val ps = initPerformanceSampler()
+      ps ! WorkTimedOut
       expectNoMsg(waitDuration)
-      mc ! WorkFailed
+      ps ! WorkFailed
       expectMsgType[Sample].workDone should be(1)
     }
 
     "resume to collect metrics once pool becomes busy again, but doesn't count old work" in {
-      val mc = initMetricsCollector()
-      mc ! partialUtilizedResult
+      val ps = initPerformanceSampler()
+      ps ! partialUtilizedResult
 
       expectMsgType[PartialUtilization]
 
-      mc ! WorkCompleted(1.millisecond)
-      mc ! WorkCompleted(1.millisecond)
+      ps ! WorkCompleted(1.millisecond)
+      ps ! WorkCompleted(1.millisecond)
 
-      mc ! fullyUtilizedResult
+      ps ! fullyUtilizedResult
 
-      mc ! WorkCompleted(1.millisecond)
+      ps ! WorkCompleted(1.millisecond)
 
       expectMsgType[Sample].workDone should be(1)
 
     }
 
     "reset counter when pool size changed" in {
-      val mc = initMetricsCollector()
+      val ps = initPerformanceSampler()
 
-      mc ! WorkCompleted(1.millisecond)
-      mc ! WorkCompleted(1.millisecond)
+      ps ! WorkCompleted(1.millisecond)
+      ps ! WorkCompleted(1.millisecond)
       expectMsgType[Sample].workDone should be(2)
 
-      mc ! PoolSize(12)
-      mc ! WorkCompleted(1.millisecond)
+      ps ! PoolSize(12)
+      ps ! WorkCompleted(1.millisecond)
 
       val sample = expectMsgType[Sample]
       sample.workDone should be(1)
@@ -94,33 +95,33 @@ class MetricsCollectorSpec extends SpecWithActorSystem with MockitoSugar with Ev
     }
 
     "register pool size when resting" in {
-      val mc = initMetricsCollector()
+      val ps = initPerformanceSampler()
 
-      mc ! partialUtilizedResult
+      ps ! partialUtilizedResult
       expectMsgType[PartialUtilization]
 
-      mc ! PoolSize(15)
-      mc ! fullyUtilizedResult
-      mc ! WorkCompleted(1.millisecond)
+      ps ! PoolSize(15)
+      ps ! fullyUtilizedResult
+      ps ! WorkCompleted(1.millisecond)
       expectMsgType[Sample].poolSize should be(15)
 
     }
 
     "continue counting when sample duration not long enough" in {
-      val mc = initMetricsCollector(0.99)
-      mc ! WorkCompleted(1.millisecond)
-      mc ! AddSample
+      val ps = initPerformanceSampler(0.99)
+      ps ! WorkCompleted(1.millisecond)
+      ps ! AddSample
       expectNoMsg(waitDuration / 5)
-      mc ! WorkCompleted(1.millisecond)
+      ps ! WorkCompleted(1.millisecond)
       expectMsgType[Sample].workDone should be(2)
     }
 
     "reset counting when pool size changed" in {
-      val mc = initMetricsCollector(0.99)
-      mc ! WorkCompleted(1.millisecond)
-      mc ! PoolSize(15)
+      val ps = initPerformanceSampler(0.99)
+      ps ! WorkCompleted(1.millisecond)
+      ps ! PoolSize(15)
       expectNoMsg(waitDuration / 5)
-      mc ! WorkCompleted(1.millisecond)
+      ps ! WorkCompleted(1.millisecond)
       val sample = expectMsgType[Sample]
       sample.workDone should be(1)
       sample.poolSize should be(15)
@@ -128,9 +129,9 @@ class MetricsCollectorSpec extends SpecWithActorSystem with MockitoSugar with Ev
 
     "forward metrics to metric reporter" in {
       val reporter = mock[Reporter]
-      val mc = system.actorOf(MetricsCollector.props(Some(reporter)))
+      val p = system.actorOf(MetricsCollector.props(Some(reporter)))
       val ps = PoolSize(3)
-      mc ! ps
+      p ! ps
 
       verify(reporter).report(ps)
     }
